+++ donation-service/src/main/java/com/gohelpfund/api/v1/donation_service/wallets/models/wallet/promise/PromiseWalletTransaction.java
@@ -38,7 +38,7 @@ public class PromiseWalletTransaction implements Serializable {
 
     @JsonProperty("amount")
     @Column(name = "amount")
-    private Integer amount;
+    private Double amount;
 
     @JsonIgnore
     @Column(name = "sender_promise_id")
@@ -87,11 +87,11 @@ public class PromiseWalletTransaction implements Serializable {
         this.transactionType = transactionType;
     }
 
-    public Integer getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
 
@@ -139,7 +139,7 @@ public class PromiseWalletTransaction implements Serializable {
         return this;
     }
 
-    public PromiseWalletTransaction withAmount(Integer amount){
+    public PromiseWalletTransaction withAmount(Double amount){
         this.setAmount(amount);
         return this;
     }
